Move onCheckoutCompleteError effect to CheckoutBlock
