+++ assets/js/base/context/cart-checkout/checkout/index.js
@@ -45,8 +45,8 @@ export const CheckoutProvider = ( {
 					activePaymentMethod={ initialActivePaymentMethod }
 				>
 					<ShippingDataProvider>
-						<CheckoutProcessor />
 						{ children }
+						<CheckoutProcessor />
 					</ShippingDataProvider>
 				</PaymentMethodDataProvider>
 			</BillingDataProvider>
